Fix IC bike SE showing only zeroes
The IC bike SE will stop sending notifications if an unsupported Fitness
Machine Control Point procedure is used. To prevent it, read the machine
fitness features to determine if the Elapsed Time Supported and the
Remaining Time Supported bits are set in order to determine if the
"Start or Resume" can be used.

Also, reduce the connection parameters to 100 instead of 120 to be able
to get the notifications which are sent every 500 ms.